Refactor pcapng: move Unknown to its own file
