+++ src/pcapng.rs
@@ -58,6 +58,7 @@ mod process_information;
 mod section_header;
 mod simple_packet;
 mod systemd_journal_export;
+mod unknown;
 
 pub use custom::*;
 pub use decryption_secrets::*;
@@ -69,6 +70,7 @@ pub use process_information::*;
 pub use section_header::*;
 pub use simple_packet::*;
 pub use systemd_journal_export::*;
+pub use unknown::*;
 
 trait PcapNGBlockParser<'a, En: PcapEndianness, O: 'a> {
     /// Minimum header size, in bytes
@@ -295,36 +297,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
 }
 
-/// Unknown block (magic not recognized, or not yet implemented)
-#[derive(Debug)]
-pub struct UnknownBlock<'a> {
-    /// Block type (little endian)
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub data: &'a [u8],
-    pub block_len2: u32,
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, UnknownBlock<'a>> for UnknownBlock<'a> {
-    const HDR_SZ: usize = 12;
-    const MAGIC: u32 = 0;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], UnknownBlock<'a>, E> {
-        let block = UnknownBlock {
-            block_type,
-            block_len1,
-            data: i,
-            block_len2,
-        };
-        Ok((i, block))
-    }
-}
-
 #[derive(Debug)]
 pub struct PcapNGOption<'a> {
     pub code: OptionCode,
@@ -511,16 +483,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-/// Parse an unknown block (little-endian)
-pub fn parse_unknownblock_le(i: &[u8]) -> IResult<&[u8], UnknownBlock, PcapError<&[u8]>> {
-    ng_block_parser::<UnknownBlock, PcapLE, _, _>()(i)
-}
-
-/// Parse an unknown block (big-endian)
-pub fn parse_unknownblock_be(i: &[u8]) -> IResult<&[u8], UnknownBlock, PcapError<&[u8]>> {
-    ng_block_parser::<UnknownBlock, PcapBE, _, _>()(i)
 }
 
 /// Parse any block, as little-endian
